Refactor: Simplifying names for clarity
