+++ src/Goban/callbacks.ts
