+++ src/Goban/canvas_utils.ts
