+++ src/Goban/focus_tracker.ts
